big things
reworked the organization and added an output file
